+++ Net_Core/Web.Angular/projects/gw-lib/core/logging/src/logging-profile.model.ts
@@ -1,7 +1,10 @@
+import { LogDestination } from "./log-destination.enum";
+
 export interface ILoggingProfile {
   account: string;
   className: string;
   component: string;
+  destination: LogDestination[];
   level: string;
   logDate: string;
   logSeqId: number;
@@ -12,6 +15,7 @@ export interface ILoggingProfile {
 export class LoggingProfile implements ILoggingProfile {
 	public logDate: string = '';
 	public logSeqId: number = 0;
+  public destination: LogDestination[] = [];
 
 	constructor(
     public account: string,